integration-tests/requirements.txt
behave==1.*
black==20.8b1
environs==4.*
isort
kombu==4.*
mypy
python-jose==3.*
reportportal-behave-client==1.*
requests==2.*
